tests/wpt/tests/svg/path/parsing/path-length-valid.tentative.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <title>SVG Path Properties: parsing path-length with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/paths.html#PathLengthProperty"/>
    <h:meta name="assert" content="path-length supports the full grammar 'none | &lt;number [0,∞]&gt;'."/>
  </metadata>
  <path id="target"/>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("path-length", "none");
test_valid_value("path-length", "0");
test_valid_value("path-length", "1");
test_valid_value("path-length", "0.5");
test_valid_value("path-length", "100");
test_valid_value("path-length", "3.14");
test_valid_value("path-length", "1e2", "100");
test_valid_value("path-length", "calc(1 + 2)", "calc(3)");
test_valid_value("path-length", "calc(infinity)");

  ]]></script>
</svg>
